5.3.0 - Update to alt:V v15
* Tracking Branch for v15

* Remove Dead Enums

* Remove V2 Module Snippet

* Update Config / Resource Builder to TOML Only

* change exported barrel files

* Change Exports to Imports

* upgrade item drops to v15, faster boot times

* ignore some files

* Patch CTRL+C Issue

* Fix CTRL + C Issue 100%

* Local Object, Global Object Changes

* Remove Dead Code

* Update to Global Blip API

* Virtual Entities for Text Labels

* Markers Updated to Virtual Entities

* Update World Notifications to Virtual Entities

* Fix Type Issue

* Static Peds Update

* Persistent Ammo

* Add Missing Dimension Setters for Controllers

* Remove File Transformer, Require .js Extension on Imports

* Update Changelog

* Update-Dependencies.js Refactored + Github Dependencies added (#445)

* Initial (Dependencies) Fix

* Add github dependencies for plugins.

* Update Dependencie Script.

* Update / Refactor Update-Dependencies Script

* Remove disable log

* Check if URL is valid (GitHub)

* Update Dependencie Script

* Remove logs

* fix windows startup

* Shave 300ms off startup time

* Update Transpilation Scripts. (#449)

* Remove Streaming Service

* remove altv-config

---------

+++ scripts/runtime/index.js
@@ -2,7 +2,13 @@ import { ChildProcess, spawn } from 'child_process';
 import fkill from 'fkill';
 import fs from 'fs';
 import crypto from 'crypto';
-import { globSync } from '../shared/fileHelpers.js';
+import { copySync, globSync, areKeyResourcesReady } from '../shared/fileHelpers.js';
+import { buildResources } from '../buildresource/index.js';
+import { runCoreCompiler } from '../compiler/core.js';
+import { runPluginsCompiler } from '../plugins/core.js';
+import { copyPluginFiles } from '../plugins/files.js';
+import { compileWebviewPlugins } from '../plugins/webview.js';
+import { updatePluginDependencies } from '../plugins/update-dependencies.js';
 
 const DEBUG = true;
 
@@ -10,12 +16,17 @@ const NO_SPECIAL_CHARACTERS = new RegExp(/^[ A-Za-z0-9_-]*$/gm);
 
 const ports = [7788, 'altv-server', 'altv-server.exe', 3399, 3001];
 const node = 'node';
-const npx = process.platform === 'win32' ? 'npx.cmd' : 'npx';
-const npm = process.platform === 'win32' ? 'npm.cmd' : 'npm';
-const serverBinary = process.platform === 'win32' ? 'altv-server.exe' : './altv-server';
+
+const isLinux = process.platform === 'win32' ? false : true;
+
+const npx = !isLinux ? 'npx.cmd' : 'npx';
+const npm = !isLinux ? 'npm.cmd' : 'npm';
+const serverBinary = !isLinux ? 'altv-server.exe' : './altv-server';
 
 const passedArguments = process.argv.slice(2).map((arg) => arg.replace('--', ''));
 const fileNameHashes = {};
+
+let configName = 'prod';
 
 if (NO_SPECIAL_CHARACTERS.test(process.cwd())) {
     console.warn(`Hey! A folder in your Athena path has special characters in it.`);
@@ -28,11 +39,6 @@ if (NO_SPECIAL_CHARACTERS.test(process.cwd())) {
  * This is the alt:V Server Process
  * @type {ChildProcess} */
 let lastServerProcess;
-
-/**
- * This is the Streamer Server Process
- * @type {ChildProcess} */
-let lastStreamerProcess;
 
 /**
  * This is the WebView Dev Server Process
@@ -68,6 +74,10 @@ async function sleep(ms) {
  * @param {ChildProcess} process
  */
 async function killChildProcess(process) {
+    if (!process) {
+        return;
+    }
+
     while (process.killed === false) {
         process.kill();
 
@@ -91,27 +101,9 @@ async function runFile(processName, ...args) {
     });
 }
 
-async function handleConfiguration() {
-    let configName = 'prod';
-
-    if (passedArguments.includes('dev')) {
-        configName = 'dev';
-    } else if (passedArguments.includes('devtest')) {
-        configName = 'devtest';
-    }
-
-    let promises = [];
-
-    if (!passedArguments.includes('dev')) {
-        promises.push(runFile(npx, 'vite', 'build', './src-webviews'));
-    }
-
-    promises.push(runFile(npx, 'altv-config', `./configs/${configName}.json`));
-    await runFile(node, './scripts/buildresource/index.js');
-    promises.push(
-        runFile(npx, 'altv-config', './scripts/buildresource/resource.json', './resources/core/resource.toml'),
-    );
-    return await Promise.all(promises);
+function handleConfiguration() {
+    copySync(`./configs/${configName}.toml`, `server.toml`);
+    buildResources();
 }
 
 async function handleViteDevServer() {
@@ -122,15 +114,13 @@ async function handleViteDevServer() {
     await runFile(node, './scripts/plugins/webview.js');
     lastViteServer = spawn(
         npx,
-        [
-            'vite',
-            './src-webviews',
-            '--clearScreen=false',
-            '--debug=true',
-            '--host=localhost',
-            '--port=3000'
-        ]
+        ['vite', './src-webviews', '--clearScreen', '--debug=true', '--host=localhost', '--port=3000'],
+        { stdio: 'pipe' },
     );
+
+    lastViteServer.stdout.on('data', (dat) => {
+        console.log(dat.toString());
+    });
 
     lastViteServer.once('close', (code) => {
         console.log(`Vite process exited with code ${code}`);
@@ -145,20 +135,6 @@ async function handleViteDevServer() {
     });
 }
 
-function handleStreamerProcess(shouldAutoRestart = false) {
-    if (lastStreamerProcess && !lastStreamerProcess.killed) {
-        lastStreamerProcess.kill();
-    }
-
-    lastStreamerProcess = spawn(node, ['./scripts/streamer/dist/index.js'], { stdio: 'inherit' });
-    lastStreamerProcess.once('close', (code) => {
-        console.log(`Streamer process exited with code ${code}`);
-        if (shouldAutoRestart) {
-            handleStreamerProcess(shouldAutoRestart);
-        }
-    });
-}
-
 async function handleServerProcess(shouldAutoRestart = false) {
     if (lastServerProcess && !lastServerProcess.killed) {
         lastServerProcess.kill();
@@ -168,8 +144,11 @@ async function handleServerProcess(shouldAutoRestart = false) {
         await runFile('chmod', '+x', `./altv-server`);
     }
 
+    await areKeyResourcesReady();
+
     if (passedArguments.includes('cdn')) {
         lastServerProcess = spawn(serverBinary, ['--justpack'], { stdio: 'inherit' });
+
         lastServerProcess.once('exit', () => {
             const finalDestination = `${process.cwd()}/cdn_upload`.replace(/\\/g, '/');
             console.log(`Files were packed for a CDN. ${finalDestination}`);
@@ -179,10 +158,10 @@ async function handleServerProcess(shouldAutoRestart = false) {
         lastServerProcess = spawn(serverBinary, { stdio: 'inherit' });
     }
 
-    lastServerProcess.once('close', (code) => {
+    lastServerProcess.once('close', async (code) => {
         console.log(`Server process exited with code ${code}`);
         if (shouldAutoRestart) {
-            handleServerProcess(shouldAutoRestart);
+            await handleServerProcess(shouldAutoRestart);
         }
     });
 }
@@ -231,30 +210,30 @@ async function refreshFileWatching() {
     }
 }
 
+/**
+ *
+ * @returns {boolean}
+ */
 async function coreBuildProcess() {
-    const start = Date.now();
+    const timer = createExecTime('>>> Core Build Time');
 
-    const coreCompilerTime = createExecTime('>>> core-compiler');
-    await runFile(node, './scripts/compiler/core');
-    coreCompilerTime.stop();
-
-    const pluginBuildTime = createExecTime('>>> core-plugins');
-    await runFile(node, './scripts/plugins/core');
-    pluginBuildTime.stop();
-
-    const mixedTime = createExecTime('>>> plugin webview, tranform, files');
-    const promises = [
-        runFile(node, './scripts/plugins/webview'),
-        runFile(node, './scripts/transform/index'),
-        runFile(node, './scripts/plugins/files'),
-    ];
+    const promises = [runCoreCompiler(), runPluginsCompiler(), compileWebviewPlugins(), copyPluginFiles()];
 
     await Promise.all(promises);
-    mixedTime.stop();
+    timer.stop();
+    return true;
+}
 
-    const transformTime = createExecTime('>>> transform-time');
-    await runFile(node, './scripts/transform/index');
-    transformTime.stop();
+async function compileWebViewPages() {
+    if (passedArguments.includes('dev')) {
+        configName = 'dev';
+    } else if (passedArguments.includes('devtest')) {
+        configName = 'devtest';
+    }
+
+    if (!passedArguments.includes('dev')) {
+        await runFile(npx, 'vite', 'build', './src-webviews');
+    }
 }
 
 async function devMode(firstRun = false) {
@@ -263,54 +242,45 @@ async function devMode(firstRun = false) {
         return;
     }
 
-    let promises = [];
-    promises.push(killChildProcess(lastStreamerProcess));
-    promises.push(killChildProcess(lastServerProcess));
+    await killChildProcess(lastServerProcess);
 
-    await Promise.all(promises);
-    promises = [];
+    const didCoreBuild = await coreBuildProcess();
+    if (!didCoreBuild) {
+        return;
+    }
 
-    promises.push(coreBuildProcess());
-    promises.push(handleConfiguration());
+    await compileWebViewPages();
+    handleConfiguration();
 
-    await Promise.all(promises);
-
-    handleStreamerProcess(false);
-    handleServerProcess(false);
+    await handleServerProcess(false);
 }
 
 async function runServer() {
     const isDev = passedArguments.includes('dev');
 
-    //Update dependencies for all the things
-    const updateTime = createExecTime('>>> update-dependencies');
-    await runFile(node, './scripts/plugins/update-dependencies');
-    updateTime.stop();
+    // Await updating all the dependencies
+    const dependencieTimer = createExecTime(`>>> Update Plugin Dependencies`);
+    await updatePluginDependencies();
+    dependencieTimer.stop();
 
     if (isDev) {
         handleViteDevServer();
     }
 
     // Has to build first before building the rest.
-    const coreBuildTime = createExecTime('>>> core-build-time');
     await coreBuildProcess();
-    coreBuildTime.stop();
-
-    const configurationTime = createExecTime('>>> handle-configuration-time');
-    await handleConfiguration();
-    configurationTime.stop();
+    await compileWebViewPages();
+    handleConfiguration();
 
     if (passedArguments.includes('dev')) {
         await sleep(50);
         await devMode(true);
-        handleStreamerProcess(false);
-        handleServerProcess(false);
+        await handleServerProcess(false);
         return;
     }
 
     await sleep(50);
-    handleStreamerProcess(true);
-    handleServerProcess(true);
+    await handleServerProcess(true);
 }
 
 if (passedArguments.includes('start')) {
